+++ ChangeLog.rst
@@ -9,6 +9,12 @@ Note worthy changes
   password was changed", including information on user agent / IP address from where the change
   originated, will be emailed.
 
+- Google: Starting from 0.52.0, the ``id_token`` is being used for extracting
+  user information.  To accommodate for scenario's where django-allauth is used
+  in contexts where the ``id_token`` is not posted, the provider now looks up
+  the required information from the ``/userinfo`` endpoint based on the access
+  token if the ``id_token`` is absent.
+
 
 Security notice
 ---------------